requirements.txt
fastapi==0.115.8
httpx==0.28.1
matplotlib==3.9.4
numpy==2.0.2
openai==1.61.0
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.37
uvicorn==0.34.0
# PyMySQL==1.1.1
# mysql-connector-python==9.2.0